Avoid double negation in phrase.
To me it seems like double negation here is harder to understand. And reader is mostly interested in things being used and not the things being unused. But also mentioning that they are both used and ignored implies that you need to put elements corresponding to each attachment and not just used ones.