vk.xml: Include `vk_platform` type-header from `vulkan/vk_platform.h` (#1538)
It is generally expected that "top level" includes are relative to the
`./include` directory in this repo.  The video headers, sitting at the
root of the `./include` folder in `./include/vk_video/vulkan_video_*.h`,
are listed as `vk_video/vulkan_video_*.h` in `vk.xml`.  Likewise the
`vk_platform` type-header is supposed to be listed as
`vulkan/vk_platform.h` to match its relative location under `./include`.