Revert change from PR #1538 (#1610)
This caused some issues with search paths per #1573. There will probably
be a downstream change to move the vk_video/ headers under
vulkan/vk_video instead of in a parallel directory. We are not entirely
settled on how to move this forward at the moment and there's a mix of
philosophical correctness and concrete build issues being discussed in