xml/vk: Reference underlying type instead of typedef in `structextends` (#1746)
Thus far `structextends` has always noted the name of the struct it
extends directly, not one of the aliases redirecting to it.  Our
generator gets "confused" by this because we haven't taught it to
perform an alias lookup here, and it seems cleaner to solve it in vk.xml
directly.  After all, many more structs that were stabilized in 1.3 had
`structextends` references to it changed to drop the (`KHR`) extension
suffix.