Remove unused VkPrivateDataSlotCreateFlagBits{,EXT} type
As with other corresponding Flags types with no bits defined yet, this
type is not usable in the API, contains no bits, and should not be
included. This follows the template of other empty Flags types by
removing the dependencies on the FlagBits type until it is needed by
some extension which reserves a bit.

Note relationship to #1738 - as with that example, the \<enum> type
definitions remain in vk.xml, but with no dependency by any core feature
or extension on them, are never reached by output generators.
Consequently the headers, refpages, etc. no longer refer to the FlagBits
type.