Fix XML tagging of VkShaderModuleCreateInfo and add explicit VU
VkShaderModuleCreateInfo appears both as an explicit parameter, and in
the pNext chain of VkPipelineShaderStageCreateInfo, The validity
generator doesn't understand this case fully.

Proposed solution is to add an explicit VU to vkCreateShaderModule
expressing the pNext constraint; make sure implicit VU for pNext are
suppressed; and properly tag the structextends attribute for the single
structure extending VkShaderModuleCreateInfo.