GitHub CI: Regenerate and build-test Rust Vulkan bindings (Ash crate)
_This complements a similar change done to the GitLab CI script._

Rust's Vulkan bindings, provided by [the Ash crate], rely on rather
strict rules/conventions in vk.xml or otherwise fail to generate.  In
order to provide more insight into the effect of spec changes to this
project, and hopefully catch mishaps earlier (ie. naming rules for enum
variants with respect to the typename of the enum) regenerate and
build-test the Rust crate in the CI.

Since Rust bindings are not part of the official Khronos Vulkan SDK /
release this check is currently allowed to fail.  Over time this task
can be expanded to notify Ash maintainers when PRs merged to `main`
break our build, so that we can resolve it quickly or request spec
changes before they are released hence set-in-stone.  Additionally
release managers can interpret build failures like [enum variant name
mismatches] and address them on their own, or for the time being ping
the Ash maintainers (me - @MarijnS95) to bring failures to our attention
or request clarification/assistance.

[the Ash crate]: https://github.com/ash-rs/ash/
[enum variant name mismatches]: https://github.com/Traverse-Research/ash/blob/c4f1c053ea36e4de91094c63109eeda677ca3ecf/generator/src/lib.rs#L1207-L1210