Clarify DebugUtils object support
The DebugUtils commands that allow a developer to associate data
with a given object handle were unclear in a number of places.
In particular, if the object passed in was an instance-level or
physical-device-level object, what were the restrictions.
This language attempts to rectify that.

Associated Issues: #1668

NOTE: This replaces PR #1901 which was submitted to resolve this
issue but appears to have gone stale.  But @pixelcluster deserves
kudos for attempting to get this fixed since it lead to many
good discussions with @krOoze as well.