Add a spec build test
This change adds a test spec filled mostly with Lorem Ipsum that can be
built in various versions and extensions.  The build results are checked
against expectations to ensure the spec output is as expected.

This is primarily a test of the asciidoc extensions that have been
implemented in this repository whose functionality is not automatically
tested.