Replace <feature number> attribute with <feature depends>
Initially this PR only updates the XML schema and vk.xml.

The generator scripts will need to be modified to do a topological sort
on the dependencies, and downstream repositories such as validation
layers and Vulkan-Hpp will need to be updated to remove or replace their
usage of the old attribute.