Fix the CullPrimitiveEXT VUID's (#2475)
* Add new VUID's to address validation needs of CullPrimitiveEXT. It can
be used as part of a structure or as a array of boolean. Corresponding
VUID changes are also needed for PrimitiveID, Layer, ViewPortIndex,
PrimitiveShadingRateEXT

* Update chapters/interfaces.adoc

Update the validation rules. Your comments look good. I will add it for rest of the builtins

Co-authored-by: Hans-Kristian Arntzen <post@arntzen-software.no>

* Update VUID's for builtins which are associated with PerPrimitiveEXT storage qualifier

Update the VUID's for builtins Layer, PrimitiveId, PrimitiveShadingRateKHR, ViewPortIndex similar to CullPrimitiveEXT as they can be part of a block or can be individual arrays for MeshEXT execution model

* Apply suggestions from code review

---------

Co-authored-by: Hans-Kristian Arntzen <post@arntzen-software.no>
Co-authored-by: Daniel Koch <dgkoch@users.noreply.github.com>