Create Antora spec nav.adoc from vkspec.adoc (#2589)
* First cut at generating nav.adoc from vkspec.adoc

Works, but will need to make sure {chapters}, {appendices}, {generated}
are set properly in navfile. This should be fine since for asciidoc
they're set in the Makefile, for antora they're set in the component
antora.yml.

* Fix spelling

* Generate working nav.adoc from vkspec.adoc

We cannot use ifdefs or attributes in nav files, so the conversion
script relies on tags to specify which include:: constructs are part of
the Antora navigation (all of them except Vulkan SC), and converts the
{chapters} and {appendices} attributes into subdirectories of the same
name.

* Clean dynamically generated navigation files