Initial pass at generating refpage aliases in Antora site build.
Strategy is to add a page-aliases attribute for generated refpages with
aliases. This directs Antora to create HTML redirects from the aliases
to the actual refpage. It is better than a 500 entry .htaccess file and
integrated with the Antora build. It would be even better to use
symlinks as the regular refpage generate process does with the
makemanaliases.py script, but that would require Antora extension work.

Replaces use of the refpage 'alias' attributes with aliases defined in
the XML (for APIs - other uses of the 'alias' attribute still supported,
but there is only one of those at present).
Does not actually get rid of the redundant 'alias' attributes yet,
however.

Minor markup cleanup including one vendor API that did not have
refpage block markup around its definition.

Minor cleanup of refpage extraction scripts.