Generate refpage aliases in Antora site build (#2610)
* Initial pass at generating refpage aliases in Antora site build.

Strategy is to add a page-aliases attribute for generated refpages with
aliases. This directs Antora to create HTML redirects from the aliases
to the actual refpage. It is better than a 500 entry .htaccess file and
integrated with the Antora build. It would be even better to use
symlinks as the regular refpage generate process does with the
makemanaliases.py script, but that would require Antora extension work.

Replaces use of the refpage 'alias' attributes with aliases defined in
the XML (for APIs - other uses of the 'alias' attribute still supported,
but there is only one of those at present).
Does not actually get rid of the redundant 'alias' attributes yet,
however.

Minor markup cleanup including one vendor API that did not have
refpage block markup around its definition.

Minor cleanup of refpage extraction scripts.

* Remove redundant 'alias' attributes from open refpage blocks (#2611)

* Remove redundant 'alias' attributes from open refpage blocks

Follows on #2610 by removing now redundant 'alias' attributes.
This information is instead extracted from the XML alias tags during
refpage generation.

No change in generated refpages, just a markup simplification. Editorial
but breaking out separately since it touches a lot of places in the spec
markup. However, this should not cause conflicts with (the great
majority of) other MRs.

* Update style guide for change in use of refpage alias attribute